VFK: Support MultiPolygons in the BUD block (fixes #11696)
MultiPolygons with more than one component may appear in the BUD block,
but never in the PAR block. We can confirm that real world data from the
Czech Office for Surveying, Mapping and Cadastre comply to that.

For backwards compatibility, using MultiPolygons in the BUD block is
controlled by a new configuration option OGR_VFK_DB_BUD_MULTI which is
turned off by default.